refactored the order and item models to cart
